requirements.txt
langchain
langchain-core
langchain-openai
langchain-mcp-adapters
python-dotenv
PyGithub
bilibili-api-python
httpx
fastapi
uvicorn
jinja2
python-multipart
wechatsogou
cachelib